+++ trunk/PPMC ORDEN N/FrequencyTable.h
@@ -15,12 +15,12 @@ namespace ppmc {
 			std::map<char, std::size_t>::iterator tableEnd();
 			std::size_t getFrecuenciaChar();
 			std::size_t getFrecuenciaEsc();
-			u_int64_t getNewBottom();
-			u_int64_t getNewTop();
+			u_int32_t getNewBottom();
+			u_int32_t getNewTop();
 			u_int32_t getTotal(){
 				return total;
 			}
-			void setUpLimits(u_int64_t, u_int64_t, char);
+			void setUpLimits(u_int32_t, u_int32_t, char);
 			bool find(char);
 			void addCharacter(char);
 			void getStringExc(std::string&);
@@ -31,8 +31,8 @@ namespace ppmc {
 			std::map<char, std::size_t> table;
 			std::size_t esc;
 			std::size_t frecuencia;
-			u_int64_t bottom;
-			u_int64_t top;
+			u_int32_t bottom;
+			u_int32_t top;
 			u_int32_t total;
 			bool primerPasada;
 	};
